Move JobController struct to job_controller package
